+++ tests/codegen/builtin_pid_tid.cpp
@@ -14,6 +14,7 @@ TEST(codegen, builtin_pid_tid_namespace)
   MockBPFtrace bpftrace;
   bpftrace.feature_ = std::make_unique<MockBPFfeature>(true);
   bpftrace.mock_in_init_pid_ns = false;
+  bpftrace.helper_check_level_ = 0;
 
   test(bpftrace, "kprobe:f { @x = pid; @y = tid }", NAME);
 }